requirements.txt
Django>=1.2
South
django-tagging
textile
docutils
# PIL installed on server?
# PIL==1.1.7

-e git+git://github.com/snowcloud/sc-utils.git#egg=sc-utils
-e hg+http://bitbucket.org/ubernostrum/django-contact-form/#egg=django-contact-form
django-pagination

boto
BeautifulSoup
-e svn+http://svn.effbot.org/public/tags/elementtree-1.3a3-20070912#egg=ElementTree
ClientForm

# sc-utils
vobject
-e git+git://github.com/jsocol/bleach.git#egg=bleach

